+++ src/helper.hpp
@@ -15,10 +15,12 @@
 
 #include <iostream>
 #include <atomic>
+#include <variant>
 
 #include <VideoMasterCppApi/helper/video.hpp>
 #include <VideoMasterCppApi/board/board.hpp>
 #include <VideoMasterCppApi/stream/sdi/sdi_stream.hpp>
+#include <VideoMasterCppApi/stream/dv/dv_stream.hpp>
 
 std::ostream& operator<<(std::ostream& os, Deltacast::Wrapper::Board& board);
 
@@ -32,6 +34,10 @@ namespace Application::Helper
 
     bool wait_for_input(Deltacast::Wrapper::BoardComponents::RxConnector& rx_connector, const std::atomic_bool& stop_is_requested);
 
+    using TechStream = std::variant<Deltacast::Wrapper::SdiStream, Deltacast::Wrapper::DvStream>;
+    TechStream open_stream(Deltacast::Wrapper::Board& board, VHD_STREAMTYPE stream_type);
+    Deltacast::Wrapper::Stream& to_base_stream(TechStream& stream);
+
     struct SdiSignalInformation
     {
         VHD_VIDEOSTANDARD video_standard;
@@ -44,12 +50,33 @@ namespace Application::Helper
                     && clock_divisor == other.clock_divisor 
                     && video_interface == other.video_interface;
         }
-        bool operator!=(const SdiSignalInformation& other) const
-        {
-            return !(*this == other);
-        }
+        bool operator!=(const SdiSignalInformation& other) const { return !(*this == other); }
     };
-    SdiSignalInformation detect_information(Deltacast::Wrapper::SdiStream& sdi_stream);
+    struct DvSignalInformation
+    {
+        unsigned int width;
+        unsigned int height;
+        bool progressive;
+        unsigned int framerate;
+        VHD_DV_CS cable_color_space;
+        VHD_DV_SAMPLING cable_sampling;
 
-    Deltacast::Wrapper::Helper::VideoCharacteristics get_video_characteristics(const SdiSignalInformation& sdi_signal_information);
+        bool operator==(const DvSignalInformation& other) const
+        {
+            return width == other.width 
+                    && height == other.height 
+                    && framerate == other.framerate
+                    && progressive == other.progressive
+                    && cable_color_space == other.cable_color_space
+                    && cable_sampling == other.cable_sampling;
+        }
+        bool operator!=(const DvSignalInformation& other) const { return !(*this == other); }
+    };
+    using SignalInformation = std::variant<SdiSignalInformation, DvSignalInformation>;
+
+    void configure_stream(TechStream& stream, const SignalInformation& signal_information);
+    void print_information(const SignalInformation& signal_information, const std::string& prefix = "");
+    SignalInformation detect_information(TechStream& stream);
+
+    Deltacast::Wrapper::Helper::VideoCharacteristics get_video_characteristics(const SignalInformation& signal_information);
 }