+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 - Complete refactoring to make use of new VideoMaster C++ Wrapper
 - Update dependency to video-viewer
 - Handle dependencies with conan 2.x and submodule instead of fetch_content
+- Update application name to reflect repository name
 
 ## Fixed